+++ webviewsrc/eventtree.ts
@@ -4,4 +4,42 @@ window.addEventListener('load', tryRun(async function() {
     // Zoom
     const contentElement = document.getElementById('eventtreecontent') as HTMLDivElement;
     enableZoom(contentElement, 0, 0);
+
+    showPictureWhenHover();
 }));
+
+function showPictureWhenHover() {
+    const eventNodes = document.getElementsByClassName('event-picture-host') as HTMLCollectionOf<HTMLDivElement>;
+    for (let i = 0; i < eventNodes.length; i++) {
+        const eventNode = eventNodes.item(i);
+        if (eventNode) {
+            showPictureWhenHoverElement(eventNode);
+        }
+    }
+}
+
+function showPictureWhenHoverElement(eventNode: HTMLDivElement) {
+    const pictureKey = eventNode.attributes.getNamedItem('picture-style-key')?.value;
+    const pictureWidthStr = eventNode.attributes.getNamedItem('picture-width')?.value;
+    if (!pictureKey || !pictureWidthStr) {
+        return;
+    }
+
+    const pictureWidth = parseInt(pictureWidthStr);
+
+    let hoverElement: HTMLDivElement | undefined = undefined;
+
+    eventNode.addEventListener('mouseenter', () => {
+        const position = eventNode.getBoundingClientRect();
+        hoverElement = document.createElement('div');
+        hoverElement.className = pictureKey;
+        hoverElement.style.position = 'absolute';
+        hoverElement.style.left = (position.left + window.scrollX - (pictureWidth - position.width) / 2) + 'px';
+        hoverElement.style.top = (position.top + position.height + window.scrollY) + 'px';
+        document.body.append(hoverElement);
+    });
+
+    eventNode.addEventListener('mouseleave', () => {
+        hoverElement?.remove();
+    });
+}